refactor : fixture 패키지 이동(application - fixture)
